refactor snackbar
